Added User Login Feature
ABOUT: this commit extends the user model and implements the log in
functionality for users. The login uses MD5 processing to securely
transmit password. An exception handler is also included as part of the
change to handle login-related exceptions.

+++ src/main/java/cc/lzhong/scalez/dao/UserDao.java
@@ -9,5 +9,5 @@ import org.apache.ibatis.annotations.Select;
 public interface UserDao {
 
     @Select("select * from User where id=#{id}")
-    public User getUserById(@Param("id") int id);
+    public User getUserById(@Param("id") Long id);
 }
